Rewrite CrBuilder to Resource::dynamic
This gives a bit more first class support for this type of object.
It also allows catching the dynamic errors in the `TryFrom` impl.
