+++ README.md
@@ -16,8 +16,9 @@ Select a version of `kube` along with the [generated k8s api types](https://gith
 
 ```toml
 [dependencies]
-kube = "0.35.1"
-kube-derive = "0.35.1"
+kube = "0.36.0"
+kube-derive = "0.36.0"
+kube-runtime = "0.36.0"
 k8s-openapi = { version = "0.8.0", default-features = false, features = ["v1_17"] }
 ```
 
@@ -188,7 +189,7 @@ or in `Cargo.toml`:
 
 ```toml
 [dependencies]
-kube = { version = "0.35.1", default-features = false, features = ["rustls-tls"] }
+kube = { version = "0.36.0", default-features = false, features = ["rustls-tls"] }
 k8s-openapi = { version = "0.8.0", default-features = false, features = ["v1_17"] }
 ```
